feat: RedisDataInitializer and refactor: PointService 와 RedisManager 분리
